pyproject.toml
[project]
name = "clothes-validator"
version = "0.1.0"
description = ""
authors = [
    {name = "lukasb27",email = "lukasball@live.co.uk"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle (>=1.7.4.2,<2.0.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
download_dataset = "clothes_validator.dataset_config:download_dataset"
delete_dataset = "clothes_validator.dataset_config:remove_dataset"
label_dataset = "clothes_validator.dataset_label:label_dataset"
create_model = "clothes_validator.create_model:main"

